requirements.txt
asgiref==3.2.3
Django==3.0
djangorestframework==3.11.0
Markdown==3.1.1
mdx-del-ins==1.0
pep562==1.0
pymdown-extensions==6.2
pytz==2017.2
six==1.10.0
sqlparse==0.3.0
